+++ app/sources/entities.js
@@ -1,0 +1,14 @@
+import config from '../config';
+import { request } from '../lib/request';
+
+export default class EntitiesSource {
+  static urlRoot = `${config.apiPath}/entities`;
+
+  static get(user) {
+    return request(this.urlRoot, {
+      method: 'GET',
+      body: ''
+    })
+    .then(result => result.json());
+  }
+}