+++ app/storages/app.js
@@ -1,23 +1,18 @@
 import { AsyncStorage } from 'react-native';
 
 export default class AppStorage {
-  static storageKey() {
-    return '';
-  }
+  static storageKey = '';
 
   static get() {
-    return new Promise((resolve, reject) => {
-      AsyncStorage.getItem(this.storageKey()).then((res) => {
-        return resolve(JSON.parse(res));
-      });
-    });
+    return AsyncStorage.getItem(this.storageKey)
+      .then(res => JSON.parse(res));
   }
 
   static set(value) {
-    AsyncStorage.setItem(this.storageKey(), JSON.stringify(value));
+    AsyncStorage.setItem(this.storageKey, JSON.stringify(value));
   }
 
   static remove() {
-    AsyncStorage.setItem(this.storageKey());
+    AsyncStorage.setItem(this.storageKey);
   }
 }