+++ app/storages/entities.js
@@ -1,7 +1,5 @@
 import AppStorage from '../storages/app';
 
 export default class EntitiesStorage extends AppStorage {
-  static storageKey() {
-    return 'entities';
-  }
+  static storageKey = 'entities';
 }